ai/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "core-ai"
version = "0.1.0"
description = "Reusable AI runtime primitives: provider protocol, messages, orchestration and provider registry"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "starlette>=0.41,<1",
    "structlog>=24.0",
    "httpx>=0.27,<1",
    "python-jose[cryptography]>=3.3",
    "google-genai>=1.0",
    "pydantic>=2.9,<3",
]

[project.optional-dependencies]
test = [
    "httpx>=0.27,<1",
]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "httpx>=0.27,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
